requirements.txt
Django==1.11.2
djangorestframework==3.6.3
pytz==2017.2
wsgiref==0.1.2
